Add files via upload
The proliferation of digital images across various platforms necessitates robust
measures to safeguard sensitive visual data from unauthorized access and
manipulation. This project aims to develop an advanced image encryption framework
leveraging state-of-the-art cryptographic techniques to ensure the confidentiality and
integrity of digital images. Through the integration of encryption algorithms and
innovative methodologies, this research endeavors to mitigate potential cyber threats
and enhance the security posture of image-based communications.